requirements.txt
opencv-python
basicsr>=1.3.3.11
facexlib>=0.2.0.3
gfpgan>=0.2.1
numpy
Pillow
torch==1.8.1
torchvision==0.9.1
scikit-image==0.18.1
Flask==1.1.1
flask-cors==3.0.8